chore: 프론트엔드 페이지 분리
