+++ frontend/src/pages/Login.tsx
@@ -1,0 +1,9 @@
+function Login() {
+  return (
+      <div>
+        Hello Login Page!
+      </div>
+  );
+}
+
+export default Login;
