+++ frontend/src/pages/MyPage.tsx
@@ -1,0 +1,9 @@
+function MyPage() {
+  return (
+      <div>
+        Hello MyPage!
+      </div>
+  );
+}
+
+export default MyPage;
